Fix #140 - Check if Brotli is supported before calling brotli_compress

+++ Classes/Generator/BrotliGenerator.php
@@ -24,6 +24,9 @@ class BrotliGenerator extends AbstractGenerator
      */
     public function generate(string $entryIdentifier, string $fileName, string $data)
     {
+        if (!$this->checkAvailable()) {
+            return;
+        }
         $contentCompress = \brotli_compress($data);
         if ($contentCompress) {
             GeneralUtility::writeFile($fileName . '.br', $contentCompress);
@@ -38,7 +41,24 @@ class BrotliGenerator extends AbstractGenerator
      */
     public function remove(string $entryIdentifier, string $fileName)
     {
+        if (!$this->checkAvailable()) {
+            return;
+        }
         $removeService = GeneralUtility::makeInstance(RemoveService::class);
         $removeService->removeFile($fileName . '.br');
+    }
+
+    /**
+     * Check if Brotli is available
+     *
+     * @return bool
+     */
+    protected function checkAvailable(): bool
+    {
+        $available = function_exists('brotli_compress');
+        if (!$available) {
+            $this->logger->error('Your server do not support Botli compression, but you enable Brotli in EXT:staticfilecache configuration');
+        }
+        return $available;
     }
 }
